backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Django backend for portfolio app with Projects and Contact APIs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "djangorestframework>=3.16.1",
    "pillow>=12.0.0",
    "ruff>=0.14.5",
]
